+++ api/api.go
@@ -1,29 +1,33 @@
 package api
 
 import (
+	"encoding/json"
 	"github.com/MinterTeam/minter-go-node/config"
 	"github.com/MinterTeam/minter-go-node/eventsdb"
 	"github.com/MinterTeam/minter-go-node/log"
 	"github.com/gorilla/mux"
 	"github.com/rs/cors"
-	"io"
-	"io/ioutil"
 	"net/http"
+	"strings"
+	"sync"
+	"sync/atomic"
 
 	"github.com/MinterTeam/minter-go-node/core/minter"
 	"github.com/MinterTeam/minter-go-node/core/state"
 	"github.com/tendermint/go-amino"
 	"github.com/tendermint/tendermint/crypto/encoding/amino"
-	"github.com/tendermint/tendermint/node"
 	rpc "github.com/tendermint/tendermint/rpc/client"
 	"strconv"
 	"time"
 )
 
 var (
-	cdc        = amino.NewCodec()
-	blockchain *minter.Blockchain
-	client     *rpc.Local
+	cdc         = amino.NewCodec()
+	blockchain  *minter.Blockchain
+	client      *rpc.Local
+	connections = int32(0)
+	limitter    = make(chan struct{}, 10)
+	cfg         = config.GetConfig()
 )
 
 func init() {
@@ -31,19 +35,24 @@ func init() {
 	eventsdb.RegisterAminoEvents(cdc)
 }
 
-func RunApi(b *minter.Blockchain, node *node.Node) {
-	client = rpc.NewLocal(node)
+func RunApi(b *minter.Blockchain, tmRPC *rpc.Local) {
+	client = tmRPC
 
 	blockchain = b
 
 	router := mux.NewRouter().StrictSlash(true)
 
-	router.HandleFunc("/api/bipVolume", wrapper(GetBipVolume)).Methods("GET")
+	stats := IpStats{
+		ips:  make(map[string]int),
+		lock: sync.Mutex{},
+	}
+
+	router.Use(RateLimit(cfg.APIPerIPLimit, cfg.APIPerIPLimitWindow, &stats))
+
 	router.HandleFunc("/api/candidates", wrapper(GetCandidates)).Methods("GET")
 	router.HandleFunc("/api/candidate/{pubkey}", wrapper(GetCandidate)).Methods("GET")
 	router.HandleFunc("/api/validators", wrapper(GetValidators)).Methods("GET")
 	router.HandleFunc("/api/balance/{address}", wrapper(GetBalance)).Methods("GET")
-	router.HandleFunc("/api/balanceWS", wrapper(GetBalanceWatcher))
 	router.HandleFunc("/api/transactionCount/{address}", wrapper(GetTransactionCount)).Methods("GET")
 	router.HandleFunc("/api/sendTransaction", wrapper(SendTransaction)).Methods("POST")
 	router.HandleFunc("/api/sendTransactionSync", wrapper(SendTransactionSync)).Methods("POST")
@@ -74,11 +83,100 @@ func RunApi(b *minter.Blockchain, node *node.Node) {
 
 func wrapper(f func(w http.ResponseWriter, r *http.Request)) func(w http.ResponseWriter, r *http.Request) {
 	return func(w http.ResponseWriter, r *http.Request) {
-		defer io.Copy(ioutil.Discard, r.Body)
-		defer r.Body.Close()
+		if atomic.LoadInt32(&connections) > int32(cfg.APISimultaneousRequests) {
+			w.Header().Set("Content-Type", "application/json; charset=UTF-8")
+			w.WriteHeader(http.StatusTooManyRequests)
+			json.NewEncoder(w).Encode(Response{
+				Code: http.StatusTooManyRequests,
+				Log:  "Too many requests",
+			})
+			return
+		}
+
+		atomic.AddInt32(&connections, 1)
+		limitter <- struct{}{}
+
+		defer func() {
+			log.With("module", "api").Info("Served API request", "req", r.RequestURI)
+			<-limitter
+			atomic.AddInt32(&connections, -1)
+		}()
 
 		f(w, r)
 	}
+}
+
+type IpStats struct {
+	ips  map[string]int
+	lock sync.Mutex
+}
+
+func (s *IpStats) Reset() {
+	s.lock.Lock()
+	defer s.lock.Unlock()
+
+	s.ips = make(map[string]int)
+}
+
+func (s *IpStats) Add(identifier string, count int) int {
+	s.lock.Lock()
+	defer s.lock.Unlock()
+
+	s.ips[identifier] += count
+
+	return s.ips[identifier]
+}
+
+type Stats interface {
+	// Reset will reset the map.
+	Reset()
+
+	// Add would add "count" to the map at the key of "identifier",
+	// and returns an int which is the total count of the value
+	// at that key.
+	Add(identifier string, count int) int
+}
+
+func RateLimit(limit int, window time.Duration, stats Stats) func(next http.Handler) http.Handler {
+	var windowStart time.Time
+
+	// Clear the rate limit stats after each window.
+	ticker := time.NewTicker(window)
+	go func() {
+		windowStart = time.Now()
+
+		for range ticker.C {
+			windowStart = time.Now()
+			stats.Reset()
+		}
+	}()
+
+	return func(next http.Handler) http.Handler {
+		h := func(w http.ResponseWriter, r *http.Request) {
+			value := int(stats.Add(identifyRequest(r), 1))
+
+			XRateLimitRemaining := limit - value
+			if XRateLimitRemaining < 0 {
+				XRateLimitRemaining = 0
+			}
+
+			w.Header().Add("X-Rate-Limit-Limit", strconv.Itoa(limit))
+			w.Header().Add("X-Rate-Limit-Remaining", strconv.Itoa(XRateLimitRemaining))
+			w.Header().Add("X-Rate-Limit-Reset", strconv.Itoa(int(window.Seconds()-time.Since(windowStart).Seconds())+1))
+
+			if value >= limit {
+				w.WriteHeader(429)
+			} else {
+				next.ServeHTTP(w, r)
+			}
+		}
+
+		return http.HandlerFunc(h)
+	}
+}
+
+func identifyRequest(r *http.Request) string {
+	return strings.Split(r.Header.Get("X-Real-IP"), ":")[0]
 }
 
 func waitForTendermint() {