+++ api/candidate.go
@@ -9,7 +9,6 @@ import (
 )
 
 func GetCandidate(w http.ResponseWriter, r *http.Request) {
-
 	vars := mux.Vars(r)
 	pubkey := types.Hex2Bytes(strings.TrimLeft(vars["pubkey"], "Mp"))