+++ cmd/minter/main.go
@@ -14,6 +14,7 @@ import (
 	"github.com/tendermint/tendermint/p2p"
 	"github.com/tendermint/tendermint/privval"
 	"github.com/tendermint/tendermint/proxy"
+	rpc "github.com/tendermint/tendermint/rpc/client"
 	"os"
 )
 
@@ -35,10 +36,15 @@ func main() {
 	app := minter.NewMinterBlockchain()
 	node := startTendermintNode(app)
 
-	app.RunRPC(node)
+	client := rpc.NewLocal(node)
+	status, _ := client.Status()
+	if status.NodeInfo.Network != genesis.Network {
+		log.Error("Different networks")
+		os.Exit(1)
+	}
 
 	if !cfg.ValidatorMode {
-		go api.RunApi(app, node)
+		go api.RunApi(app, client)
 		go gui.Run(cfg.GUIListenAddress)
 	}