+++ core/minter/minter.go
@@ -13,14 +13,11 @@ import (
 	"github.com/MinterTeam/minter-go-node/eventsdb"
 	"github.com/MinterTeam/minter-go-node/genesis"
 	"github.com/MinterTeam/minter-go-node/helpers"
-	"github.com/MinterTeam/minter-go-node/log"
 	"github.com/tendermint/go-amino"
 	abciTypes "github.com/tendermint/tendermint/abci/types"
 	"github.com/tendermint/tendermint/libs/db"
-	"github.com/tendermint/tendermint/node"
-	rpc "github.com/tendermint/tendermint/rpc/client"
 	"math/big"
-	"os"
+	"sync"
 	"sync/atomic"
 )
 
@@ -35,12 +32,16 @@ type Blockchain struct {
 	height             int64
 	rewards            *big.Int
 	validatorsStatuses map[[20]byte]int8
-	tendermintRPC      *rpc.Local
+
+	lock sync.RWMutex
 }
 
 const (
 	ValidatorPresent = 1
 	ValidatorAbsent  = 2
+
+	StateDBPrefix = "s"
+	AppDBPrefix   = "a"
 )
 
 var (
@@ -49,15 +50,15 @@ var (
 )
 
 func NewMinterBlockchain() *Blockchain {
-	ldb, err := db.NewGoLevelDB("minter", utils.GetMinterHome()+"/data")
+	ldb, err := db.NewGoLevelDB("state", utils.GetMinterHome()+"/data")
 
 	if err != nil {
 		panic(err)
 	}
 
 	blockchain = &Blockchain{
-		stateDB: db.NewPrefixDB(ldb, []byte("s")),
-		appDB:   db.NewPrefixDB(ldb, []byte("a")),
+		stateDB: db.NewPrefixDB(ldb, []byte(StateDBPrefix)),
+		appDB:   db.NewPrefixDB(ldb, []byte(AppDBPrefix)),
 	}
 
 	blockchain.updateCurrentRootHash()
@@ -70,17 +71,6 @@ func NewMinterBlockchain() *Blockchain {
 	blockchain.updateCurrentState()
 
 	return blockchain
-}
-
-func (app *Blockchain) RunRPC(node *node.Node) {
-	app.tendermintRPC = rpc.NewLocal(node)
-
-	status, _ := app.tendermintRPC.Status()
-
-	if status.NodeInfo.Network != genesis.Network {
-		log.Error("Different networks")
-		os.Exit(1)
-	}
 }
 
 func (app *Blockchain) SetOption(req abciTypes.RequestSetOption) abciTypes.ResponseSetOption {
@@ -158,7 +148,6 @@ func (app *Blockchain) BeginBlock(req abciTypes.RequestBeginBlock) abciTypes.Res
 }
 
 func (app *Blockchain) EndBlock(req abciTypes.RequestEndBlock) abciTypes.ResponseEndBlock {
-
 	var updates []abciTypes.ValidatorUpdate
 
 	stateValidators := app.stateDeliver.GetStateValidators()
@@ -176,7 +165,6 @@ func (app *Blockchain) EndBlock(req abciTypes.RequestEndBlock) abciTypes.Respons
 
 	// accumulate rewards
 	for i, val := range vals {
-
 		// skip if candidate is not present
 		if app.validatorsStatuses[val.GetAddress()] != ValidatorPresent {
 			continue
@@ -312,7 +300,6 @@ func (app *Blockchain) CheckTx(rawTx []byte) abciTypes.ResponseCheckTx {
 }
 
 func (app *Blockchain) Commit() abciTypes.ResponseCommit {
-
 	hash, _, err := app.stateDeliver.Commit(false)
 
 	if err != nil {
@@ -344,7 +331,6 @@ func (app *Blockchain) Stop() {
 }
 
 func (app *Blockchain) updateCurrentRootHash() {
-
 	// todo: make provider
 	result := app.appDB.Get([]byte("root"))
 	copy(app.rootHash[:], result)
@@ -359,14 +345,23 @@ func (app *Blockchain) updateCurrentRootHash() {
 }
 
 func (app *Blockchain) updateCurrentState() {
+	app.lock.Lock()
+	defer app.lock.Unlock()
+
 	app.stateCheck = state.NewForCheck(app.stateDeliver)
 }
 
 func (app *Blockchain) CurrentState() *state.StateDB {
+	app.lock.RLock()
+	defer app.lock.RUnlock()
+
 	return state.NewForCheck(app.stateCheck)
 }
 
 func (app *Blockchain) GetStateForHeight(height int) (*state.StateDB, error) {
+	app.lock.RLock()
+	defer app.lock.RUnlock()
+
 	return state.New(int64(height), app.stateDB)
 }
 
@@ -375,7 +370,6 @@ func (app *Blockchain) Height() int64 {
 }
 
 func (app *Blockchain) getCurrentValidators() abciTypes.ValidatorUpdates {
-
 	result := app.appDB.Get([]byte("validators"))
 
 	if len(result) == 0 {
@@ -384,7 +378,7 @@ func (app *Blockchain) getCurrentValidators() abciTypes.ValidatorUpdates {
 
 	var vals abciTypes.ValidatorUpdates
 
-	err := cdc.UnmarshalBinary(result, &vals)
+	err := cdc.UnmarshalBinaryLengthPrefixed(result, &vals)
 
 	if err != nil {
 		panic(err)
@@ -394,15 +388,11 @@ func (app *Blockchain) getCurrentValidators() abciTypes.ValidatorUpdates {
 }
 
 func (app *Blockchain) saveCurrentValidators(vals abciTypes.ValidatorUpdates) {
-	data, err := cdc.MarshalBinary(vals)
+	data, err := cdc.MarshalBinaryLengthPrefixed(vals)
 
 	if err != nil {
 		panic(err)
 	}
 
 	app.appDB.Set([]byte("validators"), data)
-}
-
-func GetBlockchain() *Blockchain {
-	return blockchain
 }