+++ core/transaction/switch_candidate_status.go
@@ -34,7 +34,6 @@ func (data SetCandidateOnData) Gas() int64 {
 }
 
 func (data SetCandidateOnData) Run(sender types.Address, tx *Transaction, context *state.StateDB, isCheck bool, rewardPool *big.Int, currentBlock int64) Response {
-
 	if !context.CoinExists(tx.GasCoin) {
 		return Response{
 			Code: code.CoinNotExists,
@@ -114,7 +113,6 @@ func (data SetCandidateOffData) Gas() int64 {
 }
 
 func (data SetCandidateOffData) Run(sender types.Address, tx *Transaction, context *state.StateDB, isCheck bool, rewardPool *big.Int, currentBlock int64) Response {
-
 	if !context.CoinExists(tx.GasCoin) {
 		return Response{
 			Code: code.CoinNotExists,