+++ core/validators/validators.go
@@ -1,13 +1,15 @@
 package validators
 
 func GetValidatorsCountForBlock(block int64) int {
-	count := 16 + (block/518400)*4
+	return 100 // only for current testnet
 
-	if count > 256 {
-		return 256
-	}
-
-	return int(count)
+	//count := 16 + (block/518400)*4
+	//
+	//if count > 256 {
+	//	return 256
+	//}
+	//
+	//return int(count)
 }
 
 func GetCandidatesCountForBlock(block int64) int {